requirements.txt
discord.py==2.3.2
protobuf==3.12.4
psycopg==3.1.18
PyNaCl==1.5.0
python-dotenv==1.0.1
spotipy==2.23.0
SQLAlchemy==2.0.28
yt_dlp==2023.7.6
